requirements.txt
fastapi
uvicorn
asyncpg
dependency_injector
pyyaml
PyJWT
python-decouple
requests
tmdbv3api